+++ tb/alu_tb.v
@@ -32,25 +32,25 @@ module alu_tb;
 		$monitor("Time = %0t | alu_ctrl = %b | a = %b | b = %b | op = %b | zero = %b",
 			  $time, alu_ctrl, a, b , op, zero);
 
-		a = 10; b = 20; alu_ctrl = add;
+		a = 10; b = 20; alu_ctrl = ALU_ADD;
 		#10
-		a = 20; b = 20; alu_ctrl = sub;
+		a = 20; b = 20; alu_ctrl = ALU_SUB;
 		#10
-		a = 32'h00000001; b = 2; alu_ctrl = sll;
+		a = 32'h00000001; b = 2; alu_ctrl = ALU_SLL;
 		#10
-		a = -10; b = 2; alu_ctrl = slt;
+		a = -10; b = 2; alu_ctrl = ALU_SLT;
 		#10
-		a = 32'hFFFFFFF0; b = 32'h00000001; alu_ctrl = sltu;
+		a = 32'hFFFFFFF0; b = 32'h00000001; alu_ctrl = ALU_SLTU;
 		#10
-		a = 32'hAAAA5555; b = 32'h5555AAAA; alu_ctrl = _xor;
+		a = 32'hAAAA5555; b = 32'h5555AAAA; alu_ctrl = ALU_XOR;
 		#10
-		a = 32'h80000000; b = 2; alu_ctrl = srl;
+		a = 32'h80000000; b = 2; alu_ctrl = ALU_SRL;
 		#10
-		a = 32'h80000000; b = 2; alu_ctrl = sra;
+		a = 32'h80000000; b = 2; alu_ctrl = ALU_SRA;
 		#10
-		a = 32'h0F0F0F0F; b = 32'hF0000000; alu_ctrl = _or;
+		a = 32'h0F0F0F0F; b = 32'hF0000000; alu_ctrl = ALU_OR;
 		#10
-		a = 32'hFF0000FF; b = 32'h0F0000F0; alu_ctrl = _and;
+		a = 32'hFF0000FF; b = 32'h0F0000F0; alu_ctrl = ALU_AND;
 		#10
 		a = 10; b = 20;	 alu_ctrl = 4'b1111;